requirements.txt
slackeventsapi
slackclient>=2.0.0
